Promoted fixed-object RA, dec, and epoch to doubles, which also increases the accuracy of routines like radec_of() that use Fixed objects internally.

+++ libastro-3.7.5/ap_as.c
@@ -18,9 +18,9 @@ ap_as (Now *np, double Mjd, double *rap, double *decp)
 	/* as -> ap */
 	zero_mem ((void *)&o, sizeof(o));
 	o.o_type = FIXED;
-	o.f_RA = (float)*rap;
-	o.f_dec = (float)*decp;
-	o.f_epoch = (float)mjd;
+	o.f_RA = *rap;
+	o.f_dec = *decp;
+	o.f_epoch = mjd;
 	memcpy ((void *)&n, (void *)np, sizeof(Now));
 	n.n_epoch = EOD;
 	obj_cir (&n, &o);
@@ -29,9 +29,9 @@ ap_as (Now *np, double Mjd, double *rap, double *decp)
 
 	/* then back to start for second order correction */
 	o.o_type = FIXED;
-	o.f_RA = (float)*rap;
-	o.f_dec = (float)*decp;
-	o.f_epoch = (float)mjd;
+	o.f_RA = *rap;
+	o.f_dec = *decp;
+	o.f_epoch = mjd;
 	memcpy ((void *)&n, (void *)np, sizeof(Now));
 	n.n_epoch = EOD;
 	obj_cir (&n, &o);
@@ -54,9 +54,9 @@ as_ap (Now *np, double Mjd, double *rap, double *decp)
 
 	zero_mem ((void *)&o, sizeof(o));
 	o.o_type = FIXED;
-	o.f_RA = (float)*rap;
-	o.f_dec = (float)*decp;
-	o.f_epoch = (float)Mjd;
+	o.f_RA = *rap;
+	o.f_dec = *decp;
+	o.f_epoch = Mjd;
 	memcpy ((void *)&n, (void *)np, sizeof(Now));
 	n.n_epoch = EOD;
 	obj_cir (&n, &o);
